% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-intermittency.R
\name{is_intermittent}
\alias{is_intermittent}
\title{Checking for flow intermittency}
\usage{
is_intermittent(time, flow, threshold = 0.001, ndays = 5, consecutive = TRUE)
}
\description{
Checking for flow intermittency
}
\examples{
\dontrun{
library(dplyr)
library(purrr)

international \%>\%
select(river, data) \%>\%
  mutate(
    is.intermittent = map_lgl(data, ~is_intermittent(.x$time, .x$discharge))
  )}
}
