% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-metrics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{baseflow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lfstat}{\code{\link[lfstat]{baseflow}}}
}}

